+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Virtual Beams Asteroid 2 Changelog
 
+<a name="Feb 13, 2018"></a>
+### Feb, 2013
+* Fix bug with login method
+
+
 <a name="Jan 30, 2018"></a>
 ### Jan 30, 2018
 * Clean code and AoT compliation sucefull